Start splitting MC info
Splitting MC info into two parts, one taking actual MC knowledge, one
calculating reco observables but with MC kinematics.
