+++ source/include/ObservableExtractionTemplates.tpp
@@ -5,7 +5,8 @@
 
 //-------------------------------------------------------------------------------------------------
 
-template <class ParticleClass> void aQGCObservablesProcessor::findJetObservables( std::vector<ParticleClass*> &jet_vector ) {
+template <class ParticleClass> 
+void aQGCObservablesProcessor::findJetObservables( std::vector<ParticleClass*> &jet_vector ) {
   VectorBosonPairFinder <ParticleClass> VBpair_finder;
   VBpair_finder.setParticleVector(jet_vector);
   VBpair_finder.findBestCandidate();
@@ -20,40 +21,7 @@ template <class ParticleClass> void aQGCObservablesProcessor::findJetObservables
 	TLorentzVector V2_jet1_tlv ( boson2_jet1->getMomentum(), 	boson2_jet1->getEnergy() );
 	TLorentzVector V2_jet2_tlv ( boson2_jet2->getMomentum(), 	boson2_jet2->getEnergy() );
   
-  // Single VB observables
-  TLorentzVector V1_tlv = V1_jet1_tlv + V1_jet2_tlv;
-  TLorentzVector V2_tlv = V2_jet1_tlv + V2_jet2_tlv;
-  
-  m_V1_m        = V1_tlv.M();
-  m_V1_pT       = V1_tlv.Pt();
-  m_V1_cosTheta = V1_tlv.CosTheta();
-  
-  m_V2_m        = V2_tlv.M();
-  m_V2_pT       = V2_tlv.Pt();
-  m_V2_cosTheta = V2_tlv.CosTheta();
-
-  // Observables in boosted single V frame
-  TLorentzVector V1_jet1_V1boosted_tlv = getBoostedVector( V1_jet1_tlv, V1_tlv );
-  TLorentzVector V2_jet1_V2boosted_tlv = getBoostedVector( V2_jet1_tlv, V2_tlv );
-  
-  m_V1_jet_absCosThetaStar = fabs( V1_jet1_V1boosted_tlv.Angle( V1_tlv.Vect() ) );
-  m_V2_jet_absCosThetaStar = fabs( V2_jet1_V2boosted_tlv.Angle( V2_tlv.Vect() ) );
-  
-  // Combined VB observables
-  TLorentzVector VV_tlv = V1_tlv + V2_tlv;
-  
-  m_VV_m      = VV_tlv.M();
-  m_VV_pT     = VV_tlv.Pt();
-  m_VV_ET     = VV_tlv.Et();
-  
-  // Observables in boosted combined VV frame
-  TLorentzVector VV_V1_VVboosted_tlv = getBoostedVector( V1_tlv, VV_tlv );
-  m_VV_V_absCosThetaStar = fabs( VV_V1_VVboosted_tlv.Angle( VV_tlv.Vect() ) );
-  
-  // Observables from recoil against VBs
-  TLorentzVector collision_system_tlv ( m_com_E*sin(0.014/2.0), 0, 0, m_com_E );
-  TLorentzVector recoil_tlv  = collision_system_tlv - VV_tlv;
-  m_recoil_m  = recoil_tlv.M();
+  this->findVectorBosonObservables( V1_jet1_tlv, V1_jet2_tlv, V2_jet1_tlv, V2_jet2_tlv  );
   
   // General jet observables
   m_min_jetE          = std::numeric_limits<float>::infinity(); // Start out with maximal minimum values -> Must be overwritten
@@ -93,7 +61,8 @@ template <class ParticleClass> void aQGCObservablesProcessor::findJetObservables
 
 //-------------------------------------------------------------------------------------------------
 
-template <class ParticleClass> void aQGCObservablesProcessor::findParticleObservables( std::vector<ParticleClass*> &particle_vector ) {
+template <class ParticleClass> 
+void aQGCObservablesProcessor::findParticleObservables( std::vector<ParticleClass*> &particle_vector ) {
   unsigned int n_particles = particle_vector.size();
   
   // Find highest energy track of the event
